Update deviceupdate.json
Updating the Account Name and Instance Name validation. It will now allow a dash(-) in the middle of the name. It will not allow a dash in the beginning or end of a name, and it will also not allow double dash(--).
Allowed example: test-account 
Not Allowed example: -test- or -test or test- or -test- or test--name